# api/public/swagger.yml
info:
  title: Inventory API
  version: v1
openapi: 3.1.0
tags:
  - name: warehouse
    description: Everything about the Warehouse
  - name: movement
    description: Start and finish the movement
paths:
  /warehouse/assign:
    post:
      tags:
        - warehouse
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
              required:
                - name
                - code
      responses:
        '200':
          description: Success operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '201':
          description: Item has been added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /warehouse/item/{code}:
    get:
      tags:
        - warehouse
      summary: Find item by code.
      description: Returns a single item.
      operationId: getItemByCode
      parameters:
        - name: code
          in: path
          description: code of item to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /movement/current:
    get:
      tags:
        - movement
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
    post:
      tags:
        - movement
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    delete:
      tags:
        - movement
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
  /movement/item:
    post:
      tags:
        - movement
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
              required:
                - code
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    delete:
      tags:
        - movement
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
components:
  schemas:
    Item:
      required:
        - code
        - name
      type: object
      properties:
        code:
          type: string
          example: 12341234
        name:
          type: string
          example: Shure sm58
    Movement:
      required:
        - hasBeenStarted
      type: object
      properties:
        hasBeenStarted:
          type: boolean
          example: false
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  requestBodies:
    Item:
      description: Item object that needs to be added to the warehouse
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Item'
        application/xml:
          schema:
            $ref: '#/components/schemas/Item'
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets